Переписан

+++ Entities/Engine.cs
@@ -3,7 +3,7 @@
 public class Engine
 {
     public int EngineId { get; set; }
-    public int EngineCapacity { get; set; }
+    public decimal EngineCapacity { get; set; }
     public int Power { get; set; }
     public string? EngineConfiguration { get; set; }
     public int Torque { get; set; }
